+++ src/cli/roam.ts
@@ -4,18 +4,20 @@ import { Command } from 'commander';
 import { createGetCommand } from './commands/get.js';
 import { createSearchCommand } from './commands/search.js';
 import { createSaveCommand } from './commands/save.js';
+import { createRefsCommand } from './commands/refs.js';
 
 const program = new Command();
 
 program
   .name('roam')
   .description('CLI for Roam Research')
-  .version('1.5.0');
+  .version('1.6.0');
 
 // Register subcommands
 program.addCommand(createGetCommand());
 program.addCommand(createSearchCommand());
 program.addCommand(createSaveCommand());
+program.addCommand(createRefsCommand());
 
 // Parse arguments
 program.parse();